% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylogenetically_weighted_connection.R
\name{distadj}
\alias{distadj}
\title{Phylogenetically weighted connectivity}
\usage{
distadj(distmat, namesvec)
}
\arguments{
\item{inputParameter1}{distmat is a 3 column matrix. Column 1 has all NTUs repeated
appropriately, column 2 has the 5 upstream and 5 downstream (if appropriate) NTUs paired with
 each corresponding NTU from column 1, and column 3 has the distance between the two
 NTUs from columns 1 and 2  \code{inputParameter1}}

\item{inputParameter2}{boolean is a matrix of 1's and 0's from the final diagnostic scripts indicating #' which NTUs are connected (1's) \code{inputParameter2}}

\item{inputParameter3}{fname is a file name for the output of the function  \code{inputParameter3}}

\item{inputParameter4}{namesvec is a vector of names for the NTUs  \code{inputParameter4}}
}
\value{
output A matrix of similarity scores for the connections between each pair of NTUs weighted by their phylogenetic distance within a +/- 5 NTU window
}
\description{
This analysis measures the similarity of connections for each pair of NTUs. The similarities are
weighted by phylogenetic distance. A profile is created for each NTU and compared, such that
matching, or near matching, connections score higher. It is necessary to have phylogenetic
distances between NTUs. This script does not provide a matrix of those values but future
versions of this package may include that script.
}
\examples{
Run script with this command:  sim_metric_csv(boolean,distmat,namesvec, fname)
}
\keyword{keywords}
