% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_false_pos.R
\name{s_matrixacf2}
\alias{s_matrixacf2}
\title{Conservative removal of potential false positives}
\usage{
s_matrixacf2(matrix, boolean, fn)
}
\arguments{
\item{inputParameter1}{matrix of NTU abundances in rows and sample dates in columns \code{inputParameter1}}

\item{inputParameter2}{boolean is a matrix of 1's and 0's from the previous linear model code indicating which NTUs are connected (1's) \code{inputParameter2}}

\item{inputParameter3}{fn is a file name for the output of the first function  \code{inputParameter3}}

\item{inputParameter4}{fn2 is a file name for the output of the second function  \code{inputParameter4}}

\item{inputParameter5}{X is the number of connections at the cutoff  \code{inputParameter5}}
}
\value{
output A boolean matrix of connections that fail (1’s). Results are in the upper triangle. Need to subtract this matrix from the matrix of connections.
}
\description{
Initial data screening suggested that most data did not need to be differenced seasonally first.
Simulations show that structure in residuals for one NTU only does not affect type I error.
However, structure in two NTU’s can greatly affect type I error. Check autocorrelation
functions for undifferenced modified data (as before). Look at row and column sums for each
NTU. Most NTUs have only a few acf’s with structure. Most of the acf’s with structure are
found in the top 5% of NTUs (sorted by the number of problem acfs). If both NTUs are in the
upper 5%, their type I error rate is probably elevated so they should be eliminated. Others are
more likely to be OK.
This is a two part script with commands in between to run manually. Determine an
appropriate cutoff (I used 5%) and then use that result in the second script. Non-automated
script is marked by comments below.
}
\examples{
Run diagnostic with this command:  s_matrixacf2(matrix=, boolean=,fn=)
}
\keyword{keywords}
