% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linmodel.R
\name{s_matrixlm}
\alias{s_matrixlm}
\title{Checks for non-zero slope in linear models of connected NTUs}
\usage{
s_matrixlm(matrix, boolean, fn)
}
\arguments{
\item{inputParameter1}{matrix of NTU abundances in rows and sample dates in columns \code{inputParameter1}}

\item{inputParameter2}{boolean is a matrix of 1's and 0's from the LSA code indicating which NTUs are connected (1's) \code{inputParameter2}}

\item{inputParameter3}{fn is a file name for the output  \code{inputParameter3}}
}
\value{
output A boolean matrix of connections that pass. Results are in the upper triangle.
}
\description{
For potential network connections between two nodal taxonomic units (NTUs),
verifies that the slope of a linear model is non-zero by ensuring that
the p value for the line is greater than 0.05
}
\examples{
s_matrixlm(matrix=deep_popcounts, boolean=combSPdeep,fn="lmconfirmation_200.csv")
}
\keyword{keywords}
